+++ packages/jest-config/src/normalize.ts
@@ -915,6 +915,7 @@ export default async function normalize(
       case 'openHandlesTimeout':
       case 'outputFile':
       case 'passWithNoTests':
+      case 'randomize':
       case 'replname':
       case 'resetMocks':
       case 'resetModules':
@@ -1028,11 +1029,14 @@ export default async function normalize(
     newOptions.onlyChanged = newOptions.watch;
   }
 
-  newOptions.showSeed = newOptions.showSeed || argv.showSeed;
+  newOptions.randomize = newOptions.randomize || argv.randomize;
+
+  newOptions.showSeed =
+    newOptions.randomize || newOptions.showSeed || argv.showSeed;
 
   const upperBoundSeedValue = 2 ** 31;
 
-  // xoroshiro128plus is used in v8 and is used here (at time of writing)
+  // bounds are determined by xoroshiro128plus which is used in v8 and is used here (at time of writing)
   newOptions.seed =
     argv.seed ??
     Math.floor((2 ** 32 - 1) * Math.random() - upperBoundSeedValue);
@@ -1042,7 +1046,7 @@ export default async function normalize(
   ) {
     throw new ValidationError(
       'Validation Error',
-      `seed value must be between \`-0x80000000\` and \`0x7fffffff\` inclusive - is ${newOptions.seed}`,
+      `seed value must be between \`-0x80000000\` and \`0x7fffffff\` inclusive - instead it is ${newOptions.seed}`,
     );
   }